Conflicts:

    * recipes-debian/openssh/openssh_debian.bb

      This merge commit merge the two options:
        openssh: Change libexecdir in configuration
        76629c758d0167c9a997fb284820e81e7ec0a5ae
        add --libexecdir=/ to EXTRA_OECONF

        libpam: Handle selinux option by DISTRO_FEATURES
        d347eb69fdd6d9656c53c97578dcec98bad1b101
        enable/disable selinux by using DISTRO_FEATURES

    * recipes-debian/nfs-utils/nfs-utils_debian.bb

      This merge commit will remove ipv6 from DISTRO_FEATURES,
        We're requires limit amount feature in DISTRO_FEATURES.

    * recipes-debian/systemd/systemd_debian.bb
    * recipes-debian/xorg-driver/xf86-video-intel_debian.bb

      Solving the conflicts of systemd and xf86-video-intel
      by replace base_contains with bb.utils.contains
      base on commit: 823852a4237e145ae3c72bc722ee66562ef062e2

+++ recipes-debian/git/git_debian.bb
@@ -10,6 +10,7 @@ HOMEPAGE = "http://git-scm.com"
 PR = "r0"
 
 inherit debian-package
+PV = "2.1.4"
 
 LICENSE = "GPLv2 & LGPLv2.1+"
 LIC_FILES_CHKSUM = " \
@@ -159,8 +160,10 @@ do_install () {
 	rm ${D}${exec_prefix}/lib/perl-native/perl/*/Error.pm
 
 	# man3 was installed in wrong location
-	mv ${D}${prefix}/man/man3 ${D}${mandir}/
-	rm -rf ${D}${prefix}/man
+	if [ -e ${D}${prefix}/man/man3 ]; then
+		mv ${D}${prefix}/man/man3 ${D}${mandir}/
+		rm -rf ${D}${prefix}/man
+	fi
 }
 
 perl_native_fixup () {
